move matplotlib related common utility functions to _engine_utils
