+++ nilearn/plotting/html_connectome.py
@@ -9,7 +9,7 @@ from scipy import sparse
 from nilearn import DEFAULT_DIVERGING_CMAP
 from nilearn._utils.html_document import HTMLDocument
 from nilearn.datasets import fetch_surf_fsaverage
-from nilearn.plotting._utils import colorscale, to_color_strings
+from nilearn.plotting._engine_utils import colorscale, to_color_strings
 from nilearn.plotting.js_plotting_utils import (
     add_js_lib,
     encode,
