+++ nilearn/plotting/html_stat_map.py
@@ -21,7 +21,7 @@ from nilearn._utils.niimg import safe_get_data
 from nilearn._utils.param_validation import check_threshold
 from nilearn.datasets import load_mni152_template
 from nilearn.image import get_data, new_img_like, reorder_img, resample_to_img
-from nilearn.plotting._utils import colorscale
+from nilearn.plotting._engine_utils import colorscale
 from nilearn.plotting.find_cuts import find_xyz_cut_coords
 from nilearn.plotting.img_plotting import load_anat
 from nilearn.plotting.js_plotting_utils import get_html_template
