+++ nilearn/plotting/surface/_matplotlib_backend.py
@@ -15,11 +15,11 @@ from nilearn._utils import compare_version
 from nilearn._utils.logger import find_stack_level
 from nilearn.image import get_data
 from nilearn.plotting import cm
+from nilearn.plotting._engine_utils import to_color_strings
 from nilearn.plotting._utils import (
     get_cbar_ticks,
     get_colorbar_and_data_ranges,
     save_figure_if_needed,
-    to_color_strings,
 )
 from nilearn.plotting.cm import mix_colormaps
 from nilearn.plotting.surface._utils import (
